fix(express): fix cat router
